% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_class.R
\docType{methods}
\name{cs_unfold-methods}
\alias{cs_unfold-methods}
\alias{cs_unfold}
\alias{cs_unfold,Tensor-method}
\title{Tensor Column Space Unfolding}
\usage{
cs_unfold(tnsr, m)

\S4method{cs_unfold}{Tensor}(tnsr, m = NULL)
}
\arguments{
\item{tnsr}{Tensor instance}

\item{m}{mode to be unfolded on}
}
\description{
Tensor Column Space Unfolding
}
\details{
\code{cs_unfold(tnsr,m=NULL)}
}
