% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricks.R
\name{normalize}
\alias{normalize}
\title{Normalization}
\usage{
normalize(x)
}
\arguments{
\item{x}{vector, the numerical vector need to be normalized}
}
\value{
the normalized vector
}
\description{
Normalize the vector to have L2 norm 1
}
